Switched up storage so I store length of deletion string instead of the actual string for a nucleotide identifier
