+++ storage/test/integration/variant/service/test_MLSTQueryService.py
@@ -3,7 +3,7 @@ import math
 import pandas as pd
 import pytest
 
-from storage.variant.model import Sample
+from storage.variant.model.db import Sample
 from storage.variant.model.QueryFeatureMLST import QueryFeatureMLST
 from storage.variant.service.MLSTQueryService import MLSTQueryService
