+++ storage/test/integration/variant/service/test_MutationQueryService.py
@@ -3,7 +3,7 @@ import math
 import pandas as pd
 import pytest
 
-from storage.variant.model import Sample
+from storage.variant.model.db import Sample
 from storage.variant.model.QueryFeatureMutation import QueryFeatureMutation
 from storage.variant.service.MutationQueryService import MutationQueryService
