+++ storage/variant/model/QueryFeatureMutation.py
@@ -1,3 +1,7 @@
+from __future__ import annotations
+from typing import Union, Tuple
+import re
+
 from storage.variant.model import NUCLEOTIDE_UNKNOWN
 from storage.variant.model.QueryFeature import QueryFeature
 
@@ -68,3 +72,50 @@ class QueryFeatureMutation(QueryFeature):
             self._ref,
             NUCLEOTIDE_UNKNOWN
         ]))
+
+    @classmethod
+    def convert_deletion(cls, deletion: Union[str, int]) -> int:
+        if isinstance(deletion, str):
+            if re.match('^\d+$', deletion):
+               deletion = int(deletion)
+            else:
+                if not set(deletion).issubset({'A','T','C','G'}):
+                    raise Exception('Deletion must either be an integer or a string with alphabet {A,T,C,G}'
+                                    f': {deletion}')
+                deletion = len(deletion)
+        elif isinstance(deletion, int):
+            if deletion < 0:
+                raise Exception(f'ref=[{deletion}] must be a non-negative integer')
+        else:
+            raise Exception(f'ref=[{deletion}] must be either a string or a non-negative integer')
+
+        return deletion
+
+    @classmethod
+    def from_spdi(cls, spdi: str) -> Tuple[str, int, int, str]:
+        if spdi is None:
+            raise Exception('Cannot parse value spdi=None')
+
+        values = spdi.split(':')
+        if len(values) != 4:
+            raise Exception(f'Incorrect number of items for spdi=[{spdi}]')
+        else:
+            position = int(values[1])
+            deletion = cls.convert_deletion(values[2])
+
+            if position < 0:
+                raise Exception(f'Position must be non-negative: {position}')
+
+            return str(values[0]), position, deletion, str(values[3])
+
+    @classmethod
+    def to_spdi(cls, sequence_name: str, position: int, ref: Union[str, int], alt: str) -> str:
+        if position < 0:
+            raise Exception(f'Position must be non-negative: {position}')
+
+        ref = cls.convert_deletion(ref)
+        return f'{sequence_name}:{position}:{ref}:{alt}'
+
+    def standardize_feature(self) -> QueryFeatureMutation:
+        standardized_spdi = QueryFeatureMutation.to_spdi(self.scope, self.position, self.ref, self.alt)
+        return QueryFeatureMutation(standardized_spdi)
