+++ storage/variant/service/VariationService.py
@@ -9,7 +9,7 @@ from storage.variant.SampleSet import SampleSet
 from storage.variant.io.FeaturesReader import FeaturesReader
 from storage.variant.io.mutation.NucleotideFeaturesReader import NucleotideFeaturesReader
 from storage.variant.io.mutation.VariationFile import VariationFile
-from storage.variant.model import Sample, SampleNucleotideVariation, NucleotideVariantsSamples
+from storage.variant.model.db import NucleotideVariantsSamples, SampleNucleotideVariation, Sample
 from storage.variant.service import DatabaseConnection
 from storage.variant.service.FeatureService import FeatureService
 from storage.variant.service.ReferenceService import ReferenceService
